modified matrix values from int to short; Reduced execution time

+++ lcs-serial.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <string>
 
 using namespace std;
@@ -21,8 +22,8 @@ int N_LENGTH;
 int M_LENGTH;
 
 //Auxiliary tables
-int** TABLE;
-int** TRACKER;
+short** TABLE;
+short** TRACKER;
 
 
 /***********************
@@ -61,12 +62,12 @@ int initProblem(string filename)
 		fclose(myfile);
 
 		//Create auxiliary tables
-		TABLE = new int*[N_LENGTH+1];
-		TRACKER = new int*[N_LENGTH+1];
+		TABLE = new short*[N_LENGTH+1];
+		TRACKER = new short*[N_LENGTH+1];
 		
 		for(int i = 0; i < N_LENGTH+1; ++i){
-			TABLE[i] = new int[M_LENGTH+1]();
-			TRACKER[i] = new int[M_LENGTH+1]();
+			TABLE[i] = new short[M_LENGTH+1]();
+			TRACKER[i] = new short[M_LENGTH+1]();
 		}
 	}
 	else{
@@ -132,7 +133,6 @@ void printResult()
 
 int main(int argc, const char* argv[])
 {
-
 	//The problem has one input. Another file.
 	if(argc != 2){
 		printf("lcs-serial should have *1* input, it has *%d*.", argc-1);
@@ -148,6 +148,8 @@ int main(int argc, const char* argv[])
 	
 	//Print
 	printResult();
+
+
 	
 	return 0;
 }
